Api versioning added, Company services added, Auth added for swagger and token repsonse model fixed, Auth added to User and Booking services.

+++ camp-fire-api/camp-fire.API/Controllers/BookingController.cs
@@ -1,12 +1,17 @@
+using Asp.Versioning;
 using camp_fire.API.Configurations;
 using camp_fire.Application.IServices;
 using camp_fire.Application.Models;
+using camp_fire.Application.Token;
+using camp_fire.Domain.SeedWork.Exceptions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace camp_fire.API.Controllers;
 
-[Route("[controller]")]
+[ApiVersion("1.0")]
+[Route("api/v{version:apiVersion}/[controller]")]
+[Authorize]
 public class BookingController : BaseApiController
 {
     private readonly ILogger<BookingController> _logger;
@@ -20,43 +25,67 @@ public class BookingController : BaseApiController
         _bookingService = bookingService;
     }
 
-    [HttpGet]
-    [AllowAnonymous]
+    [HttpGet("myBookings")]
     public async Task<IActionResult> Get([FromQuery] GetBookingsRequest request)
     {
+        var loggedInUser = TokenProvider.GetLoggedInUser(User);
+
+        request.CreatedBy = loggedInUser.Id;
+
+        var result = await _bookingService.GetMyAsync(request);
+        return Ok(new BaseApiResult { Data = result });
+    }
+
+    [HttpGet("admin")]
+    public async Task<IActionResult> GetAll([FromQuery] GetBookingsRequest request)
+    {
+        var loggedInUser = TokenProvider.GetLoggedInUser(User);
+
+        if (!loggedInUser.IsManager)
+            throw new ApiException("You don't have permission!");
+
         var result = await _bookingService.GetAsync(request);
         return Ok(new BaseApiResult { Data = result });
     }
 
-    [HttpGet("{id}")]
-    [AllowAnonymous]
-    public async Task<IActionResult> Get(int id)
+    // [HttpGet("{id}")]
+    // [AllowAnonymous]
+    // public async Task<IActionResult> Get(int id)
+    // {
+    //     var result = await _bookingService.GetByIdAsync(id);
+    //     return Ok(new BaseApiResult { Data = result });
+    // }
+
+    [HttpPost]
+    public async Task<IActionResult> Post([FromBody] CreateBookingRequest request)
     {
-        var result = await _bookingService.GetByIdAsync(id);
+        var loggedInUser = TokenProvider.GetLoggedInUser(User);
+        request.CreatedBy = loggedInUser.Id;
+
+        var result = await _bookingService.CreateAsync(request);
         return Ok(new BaseApiResult { Data = result });
     }
 
-    [HttpPost]
-    [AllowAnonymous]
-    public async Task<IActionResult> Post([FromBody] CreateBookingRequest request)
-    {
-        var result = await _bookingService.CreateAsync(request);
-        return Ok(result);
-    }
-
     [HttpPut]
-    [AllowAnonymous]
     public async Task<IActionResult> Put([FromBody] UpdateBookingRequest request)
     {
+        var loggedInUser = TokenProvider.GetLoggedInUser(User);
+
+        request.CreatedBy = loggedInUser.Id;
+
         var result = await _bookingService.UpdateAsync(request);
 
         return Ok(new BaseApiResult { Data = result });
     }
 
     [HttpDelete]
-    [AllowAnonymous]
     public async Task<IActionResult> Delete(int id)
     {
+        var loggedInUser = TokenProvider.GetLoggedInUser(User);
+
+        if (!loggedInUser.IsManager)
+            throw new ApiException("You don't have permission for delete!");
+
         await _bookingService.DeleteAsync(id);
 
         return Ok(new BaseApiResult());
